allwinner: Add A64 ccung support
Upstream DTS for A64 SoC doesn't provide a /clocks node as Linux switched
to ccu-ng
This commit adds the necessary bits to boot on pine64 with latest DTS from
upstream.
USB is not working for now and some node aren't present in the DTS (like the
PMU, Power Management Unit).

Tested on: Pine64


git-svn-id: https://svn.freebsd.org/base/head@320612 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f