MFC: r319882, r320062, r320070, r320126 Make MAXBCACHEBUF a tunable called vfs.maxbcachebuf.
By making MAXBCACHEBUF a tunable, it can be increased to allow for
larger read/write data sizes for the NFS client.
The tunable is limited to MAXPHYS, which is currently 128K.
Making MAXPHYS a tunable or increasing its value is being discussed,
since it would be nice to support a read/write data size of 1Mbyte
for the NFS client when mounting the AmazonEFS file service.

Also, define NFS_MAXXDR as the upper bound on XDR overhead in an NFS RPC.


git-svn-id: https://svn.freebsd.org/base/stable/11@320615 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f