MFC r314454,r314455:
r314454:

Use .ALLSRC instead of RPCSRC

This is a trivial simplification in the Makefile, meant to serve as
a good example for what to do with rules like this.

r314455:

Use .ALLSRC instead of RPCSRC

This is a trivial simplification in the Makefile, meant to serve as
a good example for what to do with rules like this.


git-svn-id: https://svn.freebsd.org/base/stable/10@321274 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f