MFC r320555, r320576 (by allanjude): Add -s (serial) and -p (physpath) to diskinfo
Return the bare requested information, intended for scripting.

The serial number of a SAS/SCSI device can be returned with
'camcontrol inquiry disk -S', but there is no similar switch for SATA.

This provides a way to get this information from both SAS and SATA disks

the -s and -p flags are mutually exclusive, and cannot be used with any
other flags.


git-svn-id: https://svn.freebsd.org/base/stable/11@321927 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f