MFC r312528: Make draining a sendqueue more robust.
Add own state variable to track if a sendqueue is stopped or not.
This will prevent traffic from entering the sendqueue while it is
being destroyed.

Update drain function to wait for traffic to be transmitted before
returning when the link state is active.

Add extra checks in transmit path for stopped SQ's.

While at it:
- Use likely() for a mbuf pointer check.
- Remove redundant IFF_DRV_RUNNING check.

Sponsored by:		Mellanox Technologies


git-svn-id: https://svn.freebsd.org/base/stable/11@321997 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f